SLING-7890 fixed incorrectly reported binary size
incorrectly assumed that if a valueMap.get(binaryKey,long) returned the
length for a single InputStream and valueMap.get(binaryKey,long[]) would
return an array of lengths for an array of InputStreams that
valueMap.get(binaryKey,long[]) would convert an entry of a single
InputStream to an Array with a single value. being consistent with how
other adapters work.